Format numbers in a datatable

Format datatable numeric values with commas and no decimal places

Format datatable numeric values with commas and no decimal places

Usage

DT_formatNum(
  table,
  columns = names(purrr::keep(table$x$data, is.numeric)),
  currency = "",
  interval = 3,
  mark = ",",
  digits = 0,
  dec.mark = getOption("OutDec"),
  before = TRUE,
  zero.print = NULL,
  rows = NULL
)

DT_formatNum(
  table,
  columns = names(purrr::keep(table$x$data, is.numeric)),
  currency = "",
  interval = 3,
  mark = ",",
  digits = 0,
  dec.mark = getOption("OutDec"),
  before = TRUE,
  zero.print = NULL,
  rows = NULL
)

Arguments

table

a table object created from datatable()

columns

the indices of the columns to be formatted (can be character, numeric, logical, or a formula of the form ~ V1 + V2, which is equivalent to c('V1', 'V2'))

currency

the currency symbol

interval

put a marker after how many digits of the numbers

mark

the marker after every interval decimals in the numbers

digits

the number of decimal places to round to

dec.mark

a character to indicate the decimal point

before

whether to place the currency symbol before or after the values

zero.print

a string to specify how zeros should be formatted. Useful for when many zero values exist. If NULL, keeps zero as it is.

rows

an integer vector (starting from 1) to specify the only rows that the style applies to. By default, it's NULL, meaning all rows should be formatted. Note, formatStyle() doesn't support this argument and you should use styleRow() instead. In addition, this only works expected in the client-side processing mode, i.e., server = FALSE.

Examples

library(DT)
m = cbind(matrix(rnorm(120, 1e5, 1e6), 40), runif(40), rnorm(40, 100))
colnames(m) = head(LETTERS, ncol(m))
m
#>                  A           B           C           D         E
#>  [1,]  -953928.119  -339640.44 -1239210.49 0.481481694  98.62665
#>  [2,]  -310493.403  1252320.59   102860.98 0.416871350  99.66691
#>  [3,]  1327495.872  -922691.76 -1191300.38 0.386425585  99.28724
#>  [4,]   363545.819 -1204115.34  -414841.27 0.707624473  99.78090
#>  [5,]   498455.102  1118751.10   952740.78 0.005794738 101.39168
#>  [6,]  -510234.316  -410502.08   994667.98 0.046338329 100.77425
#>  [7,]  1324364.576  -876030.24  -616501.88 0.271178704  99.66065
#>  [8,] -1043198.213   480918.79 -1075141.68 0.899220306 100.12225
#>  [9,]   629395.758 -1013042.65  -551713.10 0.729495052  98.70697
#> [10,]  -417151.998   363210.68  1097690.31 0.450143174  99.84286
#> [11,]  1048220.289  1368745.66  1488470.13 0.469732297 100.40934
#> [12,]   177257.760  -268893.70  -488061.43 0.422988001  99.67281
#> [13,]  1082995.082 -1077082.49  -466359.47 0.673710345 101.43824
#> [14,]  1383079.875  -827067.99  -195695.36 0.590408640  99.71185
#> [15,]  1705456.289   426085.89 -1133391.59 0.884573129 100.32299
#> [16,]  -365278.314   380440.05    61307.58 0.515053632 101.23655
#> [17,]  -973756.430  -858186.17   349851.89 0.238093473 100.12475
#> [18,]   784931.705  2094506.04  -178013.98 0.906032095 100.76557
#> [19,] -1206628.483   143792.25  -371475.62 0.874662398 102.61834
#> [20,]   643695.457  -821594.06  -984985.20 0.846051171  99.07196
#> [21,]  2369504.107  -665563.96  -375786.45 0.450152073  99.70661
#> [22,]   435413.825 -1242762.11  -674573.26 0.675699260  99.23336
#> [23,]   114765.060 -2128738.81    39690.10 0.881932960 101.04359
#> [24,]   412853.560  -325798.43 -1048312.09 0.108229792  99.51309
#> [25,] -1857818.593   727266.02   -84063.50 0.417343540  99.49296
#> [26,] -1093654.993  -591681.12 -1427665.76 0.249381196  99.46229
#> [27,]  1293259.182  -459443.73 -2505408.11 0.688130287 100.14923
#> [28,]  1332625.073  -267963.86  1028218.05 0.343652647 100.29082
#> [29,]  -987278.390  -728539.62   968615.55 0.020170418  99.97918
#> [30,]   982684.361    68241.12  -423964.43 0.934830377 100.88769
#> [31,]    -8856.989  -456611.33  -294356.96 0.876656255  99.72493
#> [32,]  -541079.200   777268.59 -1753693.74 0.388389992  99.10357
#> [33,] -1642945.569   956255.33   558284.89 0.953437271 100.22819
#> [34,] -1183140.761  -665448.15  1585773.78 0.229220265 100.30493
#> [35,]  -223207.528   -31886.05 -2061412.31 0.563829033 100.34353
#> [36,] -1178452.642 -1322768.53   138621.07 0.413036639  99.57588
#> [37,]  -301144.878  1278602.00  -119253.12 0.325700355  98.73041
#> [38,] -2059310.805  1057756.03   478970.85 0.109073614  99.78074
#> [39,]  -749145.399 -1012125.12 -1064903.74 0.624783701 101.60933
#> [40,]  -996515.446  -447918.43  -568603.61 0.729887884  99.97954

# format the columns A and C as currency, and D as percentages
datatable(m) %>% formatCurrency(c('A', 'C')) %>% formatPercentage('D', 2)

# the first two columns are Euro currency, and round column E to 3 decimal places
datatable(m) %>% formatCurrency(1:2, '\U20AC') %>% formatRound('E', 3)

# render vapor pressure with only two significant figures.
datatable(pressure) %>% formatSignif('pressure',2)

# apply CSS styles to columns
datatable(iris) %>%
  formatStyle('Sepal.Length', fontWeight = styleInterval(5, c('bold', 'weight'))) %>%
  formatStyle('Sepal.Width',
    color = styleInterval(3.4, c('red', 'white')),
    backgroundColor = styleInterval(3.4, c('yellow', 'gray'))
  )