Format numbers in a datatable

Format datatable numeric values with commas and no decimal places

Format datatable numeric values with commas and no decimal places

Usage

DT_formatNum(
  table,
  columns = names(purrr::keep(table$x$data, is.numeric)),
  currency = "",
  interval = 3,
  mark = ",",
  digits = 0,
  dec.mark = getOption("OutDec"),
  before = TRUE,
  zero.print = NULL,
  rows = NULL
)

DT_formatNum(
  table,
  columns = names(purrr::keep(table$x$data, is.numeric)),
  currency = "",
  interval = 3,
  mark = ",",
  digits = 0,
  dec.mark = getOption("OutDec"),
  before = TRUE,
  zero.print = NULL,
  rows = NULL
)

Arguments

table

a table object created from datatable()

columns

the indices of the columns to be formatted (can be character, numeric, logical, or a formula of the form ~ V1 + V2, which is equivalent to c('V1', 'V2'))

currency

the currency symbol

interval

put a marker after how many digits of the numbers

mark

the marker after every interval decimals in the numbers

digits

the number of decimal places to round to

dec.mark

a character to indicate the decimal point

before

whether to place the currency symbol before or after the values

zero.print

a string to specify how zeros should be formatted. Useful for when many zero values exist. If NULL, keeps zero as it is.

rows

an integer vector (starting from 1) to specify the only rows that the style applies to. By default, it's NULL, meaning all rows should be formatted. Note, formatStyle() doesn't support this argument and you should use styleRow() instead. In addition, this only works expected in the client-side processing mode, i.e., server = FALSE.

Examples

library(DT)
m = cbind(matrix(rnorm(120, 1e5, 1e6), 40), runif(40), rnorm(40, 100))
colnames(m) = head(LETTERS, ncol(m))
m
#>                  A            B           C          D         E
#>  [1,] -1300043.517   170034.850  2024343.34 0.97929341  99.83732
#>  [2,]   355317.055  -539123.324  1398392.76 0.01701569  99.17269
#>  [3,] -2337263.611    50035.101   848791.27 0.67344783 101.87651
#>  [4,]    94428.713  -151483.443   656224.33 0.37126988 100.76644
#>  [5,]   721552.721   544797.116  -448257.26 0.91801064 100.97996
#>  [6,]  1248411.606  2855417.575  1210534.89 0.67797809 101.32178
#>  [7,] -1721817.661   146531.380 -2512334.33 0.66515246  98.88029
#>  [8,]  -147325.302   677709.069   -55693.78 0.75604109 100.51460
#>  [9,]  -144199.607   218194.874   533889.79 0.54283715  98.49090
#> [10,]  -182705.449 -1811720.491  -281951.11 0.23928810 101.53274
#> [11,]  -453699.384   962086.482   524187.57 0.50889357 100.42915
#> [12,]   728982.042  -143236.740  1163102.00 0.41726437 100.12210
#> [13,]  2165024.895  -106087.195  1148712.62 0.72694885  98.86199
#> [14,] -1530989.402   119177.592    61897.11 0.63768555  99.44198
#> [15,]   612426.950   129560.754   586148.92 0.39640996 101.05254
#> [16,] -1763011.492   649827.542  1772882.61 0.95948261 100.67768
#> [17,]  -422012.515 -2174114.857  -254361.16 0.29865803 100.03850
#> [18,]    47398.090  2782557.184  1046347.89 0.05020117  99.64362
#> [19,]   642996.343  -261221.255  1416826.36 0.57618742 100.78284
#> [20,]  -814074.827   313355.750  -196640.02 0.21790581 100.80441
#> [21,]   568154.420  1174345.882  -287213.58 0.12585627  98.09994
#> [22,]   462951.256  -565088.249  -685432.66 0.93815269 100.93578
#> [23,] -1204543.545  1213952.419  -956736.87 0.80127513  99.69095
#> [24,]   837776.321  -145896.412  -695541.43 0.75805362 100.26307
#> [25,]  1988504.929 -1077563.309 -1656275.43 0.53256516  98.20941
#> [26,]     2554.896  -875850.616  -590537.90 0.54680477  99.21174
#> [27,]  -835847.354  1165057.320  -458541.99 0.09592650  98.86698
#> [28,]    84049.689   231670.635  -436663.33 0.38834975 100.36365
#> [29,]  -726788.954   588628.809   327127.13 0.17235189  99.71411
#> [30,] -1412399.651 -1599450.568  1078454.92 0.69072585 100.51767
#> [31,]  1035363.190 -1370736.306  -108882.65 0.67520850  99.89709
#> [32,]   276488.611   384150.344 -1299410.46 0.94629485  99.02593
#> [33,]   343685.465  1437320.413   358537.29 0.19621952 101.27067
#> [34,]  1723548.883   336696.283  -341799.45 0.96863750 100.96086
#> [35,]   212038.083  1418293.384   668599.86 0.38709628 100.76872
#> [36,]   -33997.013   623909.788  2226850.46 0.65034390 101.03593
#> [37,] -1810087.468   706748.047   524858.44 0.81459620  99.52611
#> [38,]  -179237.242    -9935.672 -1584281.53 0.07096477  98.72467
#> [39,]  -213445.978   272181.715   349401.78 0.52683032  99.69438
#> [40,]  1167307.879     9672.713  1172838.25 0.76347483 102.21177

# format the columns A and C as currency, and D as percentages
datatable(m) %>% formatCurrency(c('A', 'C')) %>% formatPercentage('D', 2)

# the first two columns are Euro currency, and round column E to 3 decimal places
datatable(m) %>% formatCurrency(1:2, '\U20AC') %>% formatRound('E', 3)

# render vapor pressure with only two significant figures.
datatable(pressure) %>% formatSignif('pressure',2)

# apply CSS styles to columns
datatable(iris) %>%
  formatStyle('Sepal.Length', fontWeight = styleInterval(5, c('bold', 'weight'))) %>%
  formatStyle('Sepal.Width',
    color = styleInterval(3.4, c('red', 'white')),
    backgroundColor = styleInterval(3.4, c('yellow', 'gray'))
  )